backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
langchain-core==0.1.0
requests==2.31.0
python-dotenv==1.0.0
aiosqlite==0.19.0